Solucionado error de base de datos y falta crear o mergear items para poder crear los toolkits.
