+++ src/main/java/acme/features/inventor/toolkit/InventorToolkitCreateService.java
@@ -1,4 +1,5 @@
-package acme.features.inventors.toolkits;
+
+package acme.features.inventor.toolkit;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -7,20 +8,26 @@ import acme.entities.toolkits.Toolkit;
 import acme.framework.components.models.Model;
 import acme.framework.controllers.Errors;
 import acme.framework.controllers.Request;
-import acme.framework.services.AbstractDeleteService;
+import acme.framework.services.AbstractCreateService;
 import acme.roles.Inventor;
 
-@Service
-public class InventorToolkitDeleteService implements AbstractDeleteService<Inventor, Toolkit>{
 
-	
+@Service
+public class InventorToolkitCreateService implements AbstractCreateService<Inventor, Toolkit> {
+
+	// Internal state ---------------------------------------------------------
+
 	@Autowired
 	protected InventorToolkitRepository repository;
+
+	// AbstractCreateService<Administrator, Toolkit> interface --------------
+
 
 	@Override
 	public boolean authorise(final Request<Toolkit> request) {
 		assert request != null;
-		return false;
+		request.getModel();
+		return true;
 	}
 
 	@Override
@@ -28,7 +35,7 @@ public class InventorToolkitDeleteService implements AbstractDeleteService<Inven
 		assert request != null;
 		assert entity != null;
 		assert errors != null;
-		
+
 		request.bind(entity, errors, "code", "title", "description", "assemblyNotes", "optionalLink", "totalPrice","draft");
 		
 	}
@@ -38,40 +45,49 @@ public class InventorToolkitDeleteService implements AbstractDeleteService<Inven
 		assert request != null;
 		assert entity != null;
 		assert model != null;
-		
+
 		request.unbind(entity, model, "code", "title", "description", "assemblyNotes", "optionalLink", "totalPrice","draft");
-		model.setAttribute("readonly", false);	
+		model.setAttribute("readonly", false);
 	}
 
 	@Override
-	public Toolkit findOne(final Request<Toolkit> request) {
+	public Toolkit instantiate(final Request<Toolkit> request) {	
 		assert request != null;
-		Toolkit res;
-		int id;
-		id = request.getModel().getInteger("id");
-		res = this.repository.findOneToolkitById(id);
+		
+		Toolkit toolkit;
+		toolkit = new Toolkit();
+		
+		toolkit.setCode("");
+		toolkit.setTitle("");
+		toolkit.setDescription("");
+		toolkit.setAssemblyNotes("");
+		toolkit.setOptionalLink("");
+		
+		Inventor inventor;
+		inventor = this.repository.findInventorById(request.getPrincipal().getActiveRoleId());
+		toolkit.setInventor(inventor);
 	
-		return res;
+		return toolkit;
+			
 	}
 
 	@Override
 	public void validate(final Request<Toolkit> request, final Toolkit entity, final Errors errors) {
-		assert request !=null;
+		assert request != null;
 		assert entity != null;
 		assert errors != null;
-		
+
 	}
 
 	@Override
-	public void delete(final Request<Toolkit> request, final Toolkit entity) {
+	public void create(final Request<Toolkit> request, final Toolkit entity) {
 		assert request != null;
 		assert entity != null;
-		
-		this.repository.delete(entity);
-		
+
+		entity.setDraft(false);
+		this.repository.save(entity);
 	}
-	
-	
-	
-	
+
+
+
 }
