+++ src/main/java/acme/features/inventor/toolkit/InventorToolkitDeleteService.java
@@ -1,5 +1,4 @@
-
-package acme.features.inventors.toolkits;
+package acme.features.inventor.toolkit;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -8,26 +7,20 @@ import acme.entities.toolkits.Toolkit;
 import acme.framework.components.models.Model;
 import acme.framework.controllers.Errors;
 import acme.framework.controllers.Request;
-import acme.framework.services.AbstractCreateService;
+import acme.framework.services.AbstractDeleteService;
 import acme.roles.Inventor;
 
-
 @Service
-public class InventorToolkitCreateService implements AbstractCreateService<Inventor, Toolkit> {
+public class InventorToolkitDeleteService implements AbstractDeleteService<Inventor, Toolkit>{
 
-	// Internal state ---------------------------------------------------------
-
+	
 	@Autowired
 	protected InventorToolkitRepository repository;
-
-	// AbstractCreateService<Administrator, Toolkit> interface --------------
-
 
 	@Override
 	public boolean authorise(final Request<Toolkit> request) {
 		assert request != null;
-		request.getModel();
-		return true;
+		return false;
 	}
 
 	@Override
@@ -35,7 +28,7 @@ public class InventorToolkitCreateService implements AbstractCreateService<Inven
 		assert request != null;
 		assert entity != null;
 		assert errors != null;
-
+		
 		request.bind(entity, errors, "code", "title", "description", "assemblyNotes", "optionalLink", "totalPrice","draft");
 		
 	}
@@ -45,49 +38,40 @@ public class InventorToolkitCreateService implements AbstractCreateService<Inven
 		assert request != null;
 		assert entity != null;
 		assert model != null;
-
+		
 		request.unbind(entity, model, "code", "title", "description", "assemblyNotes", "optionalLink", "totalPrice","draft");
-		model.setAttribute("readonly", false);
+		model.setAttribute("readonly", false);	
 	}
 
 	@Override
-	public Toolkit instantiate(final Request<Toolkit> request) {	
+	public Toolkit findOne(final Request<Toolkit> request) {
 		assert request != null;
-		
-		Toolkit toolkit;
-		toolkit = new Toolkit();
-		
-		toolkit.setCode("");
-		toolkit.setTitle("");
-		toolkit.setDescription("");
-		toolkit.setAssemblyNotes("");
-		toolkit.setOptionalLink("");
-		
-		Inventor inventor;
-		inventor = this.repository.findInventorById(request.getPrincipal().getActiveRoleId());
-		toolkit.setInventor(inventor);
+		Toolkit res;
+		int id;
+		id = request.getModel().getInteger("id");
+		res = this.repository.findOneToolkitById(id);
 	
-		return toolkit;
-			
+		return res;
 	}
 
 	@Override
 	public void validate(final Request<Toolkit> request, final Toolkit entity, final Errors errors) {
-		assert request != null;
+		assert request !=null;
 		assert entity != null;
 		assert errors != null;
-
+		
 	}
 
 	@Override
-	public void create(final Request<Toolkit> request, final Toolkit entity) {
+	public void delete(final Request<Toolkit> request, final Toolkit entity) {
 		assert request != null;
 		assert entity != null;
-
-		entity.setDraft(false);
-		this.repository.save(entity);
+		
+		this.repository.delete(entity);
+		
 	}
-
-
-
+	
+	
+	
+	
 }
