+++ src/main/java/acme/features/inventor/toolkit/InventorToolkitShowMineService.java
@@ -1,15 +1,13 @@
-package acme.features.inventors.toolkits;
+package acme.features.inventor.toolkit;
 
 import java.util.Collection;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import acme.entities.quantity.Quantity;
 import acme.entities.toolkits.Toolkit;
 import acme.framework.components.models.Model;
 import acme.framework.controllers.Request;
-import acme.framework.datatypes.Money;
 import acme.framework.entities.Principal;
 import acme.framework.services.AbstractShowService;
 import acme.roles.Inventor;
@@ -49,31 +47,25 @@ public class InventorToolkitShowMineService implements AbstractShowService<Inven
 	@Override
 	public Toolkit findOne(final Request<Toolkit> request) {
 		assert request != null;
-		
-		Toolkit result;
+		Toolkit res;
 		int id;
-		
 		id = request.getModel().getInteger("id");
-		result = this.repository.findOneToolkitById(id);
-		final Money totalPrice = this.calculateTotalPrice(result);
-		result.setTotalPrice(totalPrice);
-		this.repository.save(result);
-		
-		return result;
+		res = this.repository.findOneToolkitById(id);
+		return res;
 	}
 	
-	private Money calculateTotalPrice(final Toolkit t) {
-		final Money money = new Money();
-		money.setCurrency("EUR"); 
-		Double sum=0.;
-		final Collection<Quantity> quantities = this.repository.findItemQuantitiesOfToolkit(t.getId());
-		for(final Quantity quantity: quantities) {
-			sum+=quantity.getItem().getRetailPrice().getAmount()*quantity.getNumber();
-		}
-		money.setAmount(sum);
-		return money;
-	
-	}
+//	private Money calculateTotalPrice(final Toolkit t) {
+//		final Money money = new Money();
+//		money.setCurrency("EUR"); 
+//		Double sum=0.;
+//		final Collection<Quantity> quantities = this.repository.findItemQuantitiesOfToolkit(t.getId());
+//		for(final Quantity quantity: quantities) {
+//			sum+=quantity.getItem().getRetailPrice().getAmount()*quantity.getNumber();
+//		}
+//		money.setAmount(sum);
+//		return money;
+//	
+//	}
 
 	@Override
 	public void unbind (final Request<Toolkit> request, final Toolkit entity, final Model model) {
